#4058 - Cannot export texts containing certain characters as UIMA CAS XMI
- Rename packages in XMI support module
- Switch to modern spring factories declaration
