+++ inception/inception-io-xmi/src/main/java/de/tudarmstadt/ukp/inception/io/xmi/BinaryCasFormatSupport.java
@@ -15,24 +15,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.xmi;
+package de.tudarmstadt.ukp.inception.io.xmi;
 
 import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
+import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
 
 import org.apache.uima.analysis_engine.AnalysisEngineDescription;
 import org.apache.uima.cas.CAS;
+import org.apache.uima.collection.CollectionReaderDescription;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.dkpro.core.io.xml.InlineXmlWriter;
+import org.dkpro.core.io.bincas.BinaryCasReader;
+import org.dkpro.core.io.bincas.BinaryCasWriter;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
+import de.tudarmstadt.ukp.inception.io.xmi.config.UimaFormatsAutoConfiguration;
 
-public class UimaInlineXmlFormatSupport
+/**
+ * <p>
+ * This class is exposed as a Spring Component via
+ * {@link UimaFormatsAutoConfiguration#binaryCasFormatSupport}.
+ * </p>
+ */
+public class BinaryCasFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "dkpro-core-uima-inline-xml";
-    public static final String NAME = "Inline XML";
+    public static final String ID = "bin";
+    public static final String NAME = "UIMA binary CAS";
 
     @Override
     public String getId()
@@ -47,9 +57,29 @@ public class UimaInlineXmlFormatSupport
     }
 
     @Override
+    public boolean isReadable()
+    {
+        return true;
+    }
+
+    @Override
     public boolean isWritable()
     {
         return true;
+    }
+
+    @Override
+    public boolean isProneToInconsistencies()
+    {
+        return true;
+    }
+
+    @Override
+    public CollectionReaderDescription getReaderDescription(Project aProject,
+            TypeSystemDescription aTSD)
+        throws ResourceInitializationException
+    {
+        return createReaderDescription(BinaryCasReader.class, aTSD);
     }
 
     @Override
@@ -57,6 +87,6 @@ public class UimaInlineXmlFormatSupport
             TypeSystemDescription aTSD, CAS aCAS)
         throws ResourceInitializationException
     {
-        return createEngineDescription(InlineXmlWriter.class, aTSD);
+        return createEngineDescription(BinaryCasWriter.class, aTSD);
     }
 }
