+++ inception/inception-io-xmi/src/main/java/de/tudarmstadt/ukp/inception/io/xmi/UimaInlineXmlFormatSupport.java
@@ -15,34 +15,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.xmi;
+package de.tudarmstadt.ukp.inception.io.xmi;
 
 import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
-import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
 
 import org.apache.uima.analysis_engine.AnalysisEngineDescription;
 import org.apache.uima.cas.CAS;
-import org.apache.uima.collection.CollectionReaderDescription;
 import org.apache.uima.resource.ResourceInitializationException;
 import org.apache.uima.resource.metadata.TypeSystemDescription;
-import org.dkpro.core.io.xmi.XmiReader;
-import org.dkpro.core.io.xmi.XmiWriter;
+import org.dkpro.core.io.xml.InlineXmlWriter;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
-import de.tudarmstadt.ukp.clarin.webanno.xmi.config.UimaFormatsAutoConfiguration;
 
-/**
- * <p>
- * This class is exposed as a Spring Component via
- * {@link UimaFormatsAutoConfiguration#xmiFormatSupport}.
- * </p>
- */
-public class XmiFormatSupport
+public class UimaInlineXmlFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "xmi";
-    public static final String NAME = "UIMA CAS XMI (XML 1.0)";
+    public static final String ID = "dkpro-core-uima-inline-xml";
+    public static final String NAME = "Inline XML";
 
     @Override
     public String getId()
@@ -57,29 +47,9 @@ public class XmiFormatSupport
     }
 
     @Override
-    public boolean isReadable()
-    {
-        return true;
-    }
-
-    @Override
     public boolean isWritable()
     {
         return true;
-    }
-
-    @Override
-    public boolean isProneToInconsistencies()
-    {
-        return true;
-    }
-
-    @Override
-    public CollectionReaderDescription getReaderDescription(Project aProject,
-            TypeSystemDescription aTSD)
-        throws ResourceInitializationException
-    {
-        return createReaderDescription(XmiReader.class, XmiReader.PARAM_LENIENT, true);
     }
 
     @Override
@@ -87,6 +57,6 @@ public class XmiFormatSupport
             TypeSystemDescription aTSD, CAS aCAS)
         throws ResourceInitializationException
     {
-        return createEngineDescription(XmiWriter.class, aTSD, XmiWriter.PARAM_VERSION, "1.0");
+        return createEngineDescription(InlineXmlWriter.class, aTSD);
     }
 }
