+++ inception/inception-io-xmi/src/main/java/de/tudarmstadt/ukp/inception/io/xmi/XmiFormatSupport.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.tudarmstadt.ukp.clarin.webanno.xmi;
+package de.tudarmstadt.ukp.inception.io.xmi;
 
 import static org.apache.uima.fit.factory.AnalysisEngineFactory.createEngineDescription;
 import static org.apache.uima.fit.factory.CollectionReaderFactory.createReaderDescription;
@@ -30,19 +30,19 @@ import org.dkpro.core.io.xmi.XmiWriter;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.format.FormatSupport;
 import de.tudarmstadt.ukp.clarin.webanno.model.Project;
-import de.tudarmstadt.ukp.clarin.webanno.xmi.config.UimaFormatsAutoConfiguration;
+import de.tudarmstadt.ukp.inception.io.xmi.config.UimaFormatsAutoConfiguration;
 
 /**
  * <p>
  * This class is exposed as a Spring Component via
- * {@link UimaFormatsAutoConfiguration#xmiXml11FormatSupport}.
+ * {@link UimaFormatsAutoConfiguration#xmiFormatSupport}.
  * </p>
  */
-public class XmiXml11FormatSupport
+public class XmiFormatSupport
     implements FormatSupport
 {
-    public static final String ID = "xmi-xml1.1";
-    public static final String NAME = "UIMA CAS XMI (XML 1.1)";
+    public static final String ID = "xmi";
+    public static final String NAME = "UIMA CAS XMI (XML 1.0)";
 
     @Override
     public String getId()
@@ -87,6 +87,6 @@ public class XmiXml11FormatSupport
             TypeSystemDescription aTSD, CAS aCAS)
         throws ResourceInitializationException
     {
-        return createEngineDescription(XmiWriter.class, aTSD, XmiWriter.PARAM_VERSION, "1.1");
+        return createEngineDescription(XmiWriter.class, aTSD, XmiWriter.PARAM_VERSION, "1.0");
     }
 }
