--- inception/inception-io-xmi/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-de.tudarmstadt.ukp.clarin.webanno.xmi.config.UimaFormatsAutoConfiguration
