+++ inception/inception-io-xmi/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+de.tudarmstadt.ukp.inception.io.xmi.config.UimaFormatsAutoConfiguration
